Some parser cleanup
