+++ src/main/kotlin/com/github/imflog/schema/registry/parser/JsonSchemaParser.kt
@@ -1,11 +1,8 @@
 package com.github.imflog.schema.registry.parser
 
-import com.github.imflog.schema.registry.SchemaParsingException
-import com.github.imflog.schema.registry.SchemaType
 import com.github.imflog.schema.registry.LocalReference
-import io.confluent.kafka.schemaregistry.ParsedSchema
+import com.github.imflog.schema.registry.SchemaType
 import io.confluent.kafka.schemaregistry.client.SchemaRegistryClient
-import io.confluent.kafka.schemaregistry.client.rest.entities.SchemaReference
 import org.json.JSONObject
 import java.io.File
 
@@ -20,30 +17,18 @@ class JsonSchemaParser(
         private const val DEFS_NODE = "${"$"}defs"
     }
 
-    override fun parseSchemaFromFile(
-        subject: String,
-        schemaPath: String,
-        remoteReferences: List<SchemaReference>,
-        localReferences: List<LocalReference>
-    ): ParsedSchema {
-        val schemaContent = rootDir.resolve(schemaPath).readText()
-        val schema = resolveLocalReferences(schemaContent, localReferences)
-        return client
-            .parseSchema(schemaType.registryType, schema.toString(), remoteReferences)
-            .orElseThrow { SchemaParsingException(subject, schemaType) }
-    }
-
     // TODO:
     //  * Test local + remote
     //  * Create an unit test that use https://github.com/everit-org/json-schema to check the schema validity
-    private fun resolveLocalReferences(
+    override fun resolveLocalReferences(
+        subject: String,
         schemaContent: String,
         localReferences: List<LocalReference>
-    ): JSONObject {
+    ): String {
         val jsonObj = JSONObject(schemaContent)
         val localDefNodes = JSONObject()
         localReferences.forEach { reference -> localDefNodes.put(reference.name, JSONObject(reference.content(rootDir))) }
         jsonObj.append(DEFS_NODE, localDefNodes)
-        return jsonObj
+        return jsonObj.toString()
     }
 }
