+++ src/main/kotlin/com/github/imflog/schema/registry/parser/ProtobufSchemaParser.kt
@@ -1,11 +1,8 @@
 package com.github.imflog.schema.registry.parser
 
-import com.github.imflog.schema.registry.SchemaParsingException
-import com.github.imflog.schema.registry.SchemaType
 import com.github.imflog.schema.registry.LocalReference
-import io.confluent.kafka.schemaregistry.ParsedSchema
+import com.github.imflog.schema.registry.SchemaType
 import io.confluent.kafka.schemaregistry.client.SchemaRegistryClient
-import io.confluent.kafka.schemaregistry.client.rest.entities.SchemaReference
 import java.io.File
 
 class ProtobufSchemaParser(
@@ -15,15 +12,11 @@ class ProtobufSchemaParser(
 
     override val schemaType: SchemaType = SchemaType.PROTOBUF
 
-    override fun parseSchemaFromFile(
+    override fun resolveLocalReferences(
         subject: String,
-        schemaPath: String,
-        remoteReferences: List<SchemaReference>,
+        schemaContent: String,
         localReferences: List<LocalReference>
-    ): ParsedSchema {
-        val schemaContent = rootDir.resolve(schemaPath).readText()
-        return client
-            .parseSchema(schemaType.registryType, schemaContent, remoteReferences)
-            .orElseThrow { SchemaParsingException(subject, schemaType) }
+    ): String {
+        throw NotImplementedError("Local reference is not available for Protobuf yet")
     }
 }
