+++ src/main/kotlin/com/github/imflog/schema/registry/parser/SchemaParser.kt
@@ -1,8 +1,8 @@
 package com.github.imflog.schema.registry.parser
 
+import com.github.imflog.schema.registry.LocalReference
 import com.github.imflog.schema.registry.SchemaParsingException
 import com.github.imflog.schema.registry.SchemaType
-import com.github.imflog.schema.registry.LocalReference
 import io.confluent.kafka.schemaregistry.ParsedSchema
 import io.confluent.kafka.schemaregistry.client.SchemaRegistryClient
 import io.confluent.kafka.schemaregistry.client.rest.entities.SchemaReference
@@ -15,6 +15,10 @@ abstract class SchemaParser(
     abstract val schemaType: SchemaType
 
     companion object {
+        /**
+         * This is like a factory for our local parsers.
+         * We can afford to recreate the parser each time as it's not a long-running processes.
+         */
         fun provide(schemaType: SchemaType, client: SchemaRegistryClient, rootDir: File): SchemaParser =
             when (schemaType) {
                 SchemaType.AVRO -> AvroSchemaParser(client, rootDir)
@@ -23,17 +27,27 @@ abstract class SchemaParser(
             }
     }
 
-    /**
-     * Parses a schema from a file.
-     * Default implementation does not resolve the local references.
-     * Specific implementation can change based on the different protocols.
-     */
-    @Throws(SchemaParsingException::class)
-    abstract fun parseSchemaFromFile(
+    @Throws(SchemaParsingException::class, NotImplementedError::class)
+    fun parseSchemaFromFile(
         subject: String,
         schemaPath: String,
         remoteReferences: List<SchemaReference>,
         localReferences: List<LocalReference>,
-    ): ParsedSchema
+    ): ParsedSchema {
+        val schemaContent = rootDir.resolve(schemaPath).readText()
+        val parsedLocalSchemaString = if (localReferences.isNotEmpty()) {
+            resolveLocalReferences(subject, schemaContent, localReferences)
+        } else schemaContent
+
+        return client
+            .parseSchema(schemaType.registryType, parsedLocalSchemaString, remoteReferences)
+            .orElseThrow { SchemaParsingException(subject, schemaType) }
+    }
+
+    abstract fun resolveLocalReferences(
+        subject: String,
+        schemaContent: String,
+        localReferences: List<LocalReference>
+    ): String
 }
